+++ .setup/cpp/simStateSpace-sim-ssm-1-var-vary-dot.cpp
@@ -6,12 +6,7 @@
 #include <RcppArmadillo.h>
 // [[Rcpp::depends(RcppArmadillo)]]
 // [[Rcpp::export(.SimSSM1VARVary)]]
-Rcpp::List SimSSM1VARVary(const int n, const Rcpp::List& mu0,
-                          const Rcpp::List& sigma0_sqrt,
-                          const Rcpp::List& alpha, const Rcpp::List& beta,
-                          const Rcpp::List& psi_sqrt,
-                          const Rcpp::List& gamma_eta, const Rcpp::List& x,
-                          const int time, const int burn_in) {
+Rcpp::List SimSSM1VARVary(const int n, const Rcpp::List& mu0, const Rcpp::List& sigma0_sqrt, const Rcpp::List& alpha, const Rcpp::List& beta, const Rcpp::List& psi_sqrt, const Rcpp::List& gamma_eta, const Rcpp::List& x, const int time, const int burn_in) {
   // Step 1: Determine indices
   int total_time = time + burn_in;
   arma::vec mu0_temp = mu0[0];
@@ -33,19 +28,18 @@ Rcpp::List SimSSM1VARVary(const int n, const Rcpp::List& mu0,
     arma::mat psi_sqrt_temp = psi_sqrt[i];
     arma::mat gamma_eta_temp = gamma_eta[i];
 
-    // Step 3.2: Generate initial condition
+    // Step 3.2: Generate initial condition    
     eta.col(0) = mu0_temp + sigma0_sqrt_temp * arma::randn(num_latent_vars);
 
     // Step 3.3: Simulate state space model data using a loop
     for (int t = 1; t < total_time; t++) {
-      eta.col(t) = alpha_temp + beta_temp * eta.col(t - 1) +
-                   psi_sqrt_temp * arma::randn(num_latent_vars) +
-                   gamma_eta_temp * x_t.col(t);
+      eta.col(t) = alpha_temp + beta_temp * eta.col(t - 1) + psi_sqrt_temp * arma::randn(num_latent_vars) + gamma_eta_temp * x_t.col(t);
     }
 
     // Step 3.4: If there is a burn-in period, remove it
     if (burn_in > 0) {
       eta = eta.cols(burn_in, total_time - 1);
+      x_t = x_t.cols(burn_in, total_time - 1);
     }
 
     // Step 3.5: Create a vector of ID numbers of length time
@@ -53,9 +47,7 @@ Rcpp::List SimSSM1VARVary(const int n, const Rcpp::List& mu0,
     id.fill(i + 1);
 
     // Step 3.6: Save the transposed data matrices in a list
-    out[i] = Rcpp::List::create(
-        Rcpp::Named("y") = eta.t(), Rcpp::Named("eta") = eta.t(),
-        Rcpp::Named("time") = arma::regspace(1, time), Rcpp::Named("id") = id);
+    out[i] = Rcpp::List::create(Rcpp::Named("y") = eta.t(), Rcpp::Named("eta") = eta.t(), Rcpp::Named("x") = x_t.t(), Rcpp::Named("time") = arma::regspace(0, time - 1), Rcpp::Named("id") = id);
   }
 
   // Step 4: Return the results