+++ R/simStateSpace-sim-ssm-var-vary.R
@@ -57,7 +57,7 @@
 #'         data = rnorm(n = k * (time + burn_in)),
 #'         ncol = k
 #'       )
-#'     )
+#'    )
 #'   }
 #' )
 #'
@@ -103,7 +103,7 @@ SimSSMVARVary <- function(n,
                           gamma_eta = NULL,
                           x = NULL,
                           time = 0,
-                          burn_in) {
+                          burn_in = 0) {
   if (is.null(gamma_eta) || is.null(x)) {
     return(
       .SimSSM0VARVary(