+++ tests/testthat/test-simStateSpace-sim-ssm-var-fixed.R
@@ -17,7 +17,7 @@ lapply(
     beta <- diag(x = 0.5, nrow = k)
     psi_sqrt <- iden_sqrt
     time <- 50
-    burn_in <- 0
+    burn_in <- 10
     gamma_eta <- 0.10 * diag(k)
     x <- lapply(
       X = seq_len(n),
@@ -27,7 +27,7 @@ lapply(
             data = rnorm(n = k * (time + burn_in)),
             ncol = k
           )
-        )
+       )
       }
     )
 
@@ -45,6 +45,8 @@ lapply(
 
     Sim2Matrix(ssm, eta = TRUE)
     Sim2Matrix(ssm, eta = FALSE)
+    Sim2Matrix(ssm, eta = TRUE, long = FALSE)
+    Sim2Matrix(ssm, eta = FALSE, long = FALSE)
 
     # With covariates
     ssm <- SimSSMVARFixed(
@@ -62,6 +64,9 @@ lapply(
 
     Sim2Matrix(ssm, eta = TRUE)
     Sim2Matrix(ssm, eta = FALSE)
+    Sim2Matrix(ssm, eta = TRUE, long = FALSE)
+    Sim2Matrix(ssm, eta = FALSE, long = FALSE)
+
   },
   text = "test-simStateSpace-sim-ssm-var-fixed"
 )