+++ tests/testthat/test-simStateSpace-sim-ssm-var-vary.R
@@ -24,7 +24,7 @@ lapply(
     )
     psi_sqrt <- list(iden_sqrt)
     time <- 50
-    burn_in <- 0
+    burn_in <- 10
     gamma_eta <- list(0.10 * diag(k))
     x <- lapply(
       X = seq_len(n),
@@ -34,7 +34,7 @@ lapply(
             data = rnorm(n = k * (time + burn_in)),
             ncol = k
           )
-        )
+       )
       }
     )
 
@@ -49,8 +49,11 @@ lapply(
       time = time,
       burn_in = burn_in
     )
+
     Sim2Matrix(ssm, eta = TRUE)
     Sim2Matrix(ssm, eta = FALSE)
+    Sim2Matrix(ssm, eta = TRUE, long = FALSE)
+    Sim2Matrix(ssm, eta = FALSE, long = FALSE)
 
     # With covariates
     ssm <- SimSSMVARVary(
@@ -65,8 +68,12 @@ lapply(
       time = time,
       burn_in = burn_in
     )
+
     Sim2Matrix(ssm, eta = TRUE)
     Sim2Matrix(ssm, eta = FALSE)
+    Sim2Matrix(ssm, eta = TRUE, long = FALSE)
+    Sim2Matrix(ssm, eta = FALSE, long = FALSE)
+
   },
   text = "test-simStateSpace-sim-ssm-var-vary"
 )