+++ tests/testthat/test-simStateSpace-sim-ssm-vary.R
@@ -27,7 +27,7 @@ lapply(
     lambda <- list(iden)
     theta_sqrt <- list(chol(diag(x = 0.50, nrow = k)))
     time <- 50
-    burn_in <- 0
+    burn_in <- 10
     gamma_y <- gamma_eta <- list(0.10 * diag(k))
     x <- lapply(
       X = seq_len(n),
@@ -37,7 +37,7 @@ lapply(
             data = rnorm(n = k * (time + burn_in)),
             ncol = k
           )
-        )
+       )
       }
     )
 
@@ -59,6 +59,8 @@ lapply(
 
     Sim2Matrix(ssm, eta = TRUE)
     Sim2Matrix(ssm, eta = FALSE)
+    Sim2Matrix(ssm, eta = TRUE, long = FALSE)
+    Sim2Matrix(ssm, eta = FALSE, long = FALSE)
 
     # Type 1
     ssm <- SimSSMVary(
@@ -80,6 +82,8 @@ lapply(
 
     Sim2Matrix(ssm, eta = TRUE)
     Sim2Matrix(ssm, eta = FALSE)
+    Sim2Matrix(ssm, eta = TRUE, long = FALSE)
+    Sim2Matrix(ssm, eta = FALSE, long = FALSE)
 
     # Type 2
     ssm <- SimSSMVary(
@@ -102,6 +106,9 @@ lapply(
 
     Sim2Matrix(ssm, eta = TRUE)
     Sim2Matrix(ssm, eta = FALSE)
+    Sim2Matrix(ssm, eta = TRUE, long = FALSE)
+    Sim2Matrix(ssm, eta = FALSE, long = FALSE)
+
   },
   text = "test-simStateSpace-sim-ssm-vary"
 )